requirements.txt
AHRS==0.3.0
numpy==1.20.3
paho-mqtt==1.5.1
pandas==1.1.5
PyOpenGL==3.1.5
PyOpenGL-accelerate==3.1.5
python-dateutil==2.8.1
pytz==2021.1
scipy==1.6.3
six==1.16.0